Add Jest and related testing packages as dev dependencies
In this update, the test framework Jest along with ts-jest and @types/jest packages have been added to the project as dev dependencies. This is part of setting up a solid testing environment and strategy for the project according to best practices. It enhances the reliability of the code and encourages its maintainability through regression testing.